example/requirements.txt
nicolasflamel.mwc-pay==0.0.7
